Fix CSS not loaded because of MIME-type
The issue was due to the compojure `not-found` route.

The `not-found` route came before resources (css, img), so the
webserver was not able to find those resources (e.g. style.css)
and was returning some content payload of html type (hence the
resource file was blocked because of MIME type).